+++ CafeNet/Data/Models/Tax.cs
@@ -4,7 +4,7 @@
     {
         public long Id { get; set; }
         public string Type { get; set; }
-        public short Percent { get; set; }
+        public byte Percent { get; set; }
         public bool IsArchived { get; set; }
     }
 }
